+++ tests/log_to_journal.rs
@@ -11,6 +11,7 @@
 
 #![deny(warnings, clippy::all)]
 
+use log::kv::Value;
 use log::{Level, Log, Record};
 use pretty_assertions::assert_eq;
 
@@ -144,4 +145,38 @@ fn escaped_extra_fields() {
     );
     assert_eq!(entry["MESSAGE"], "with an escaped extra field");
     assert_eq!(entry["HALL_CHEN"], "Welt")
+}
+
+#[test]
+fn extra_record_fields() {
+    let target = journal::random_target("systemd_journal_logger/extra_record_fields");
+
+    let kvs: &[(&str, Value)] = &[
+        ("_foo", Value::from("foo")),
+        ("spam_with_eggs", Value::from(false)),
+    ];
+
+    JournalLog::with_extra_fields(vec![("EXTRA_FIELD", "foo")]).log(
+        &Record::builder()
+            .level(Level::Error)
+            .target(&target)
+            .module_path(Some(module_path!()))
+            .args(format_args!("Hello world"))
+            .key_values(&kvs)
+            .build(),
+    );
+
+    let entries = journal::read_current_process(module_path!(), &target);
+    assert_eq!(entries.len(), 1);
+    let entry = &entries[0];
+
+    assert_eq!(entry["TARGET"], target);
+    assert_eq!(
+        entry["PRIORITY"],
+        u8::from(libsystemd::logging::Priority::Error).to_string()
+    );
+    assert_eq!(entry["MESSAGE"], "Hello world");
+    assert_eq!(entry["EXTRA_FIELD"], "foo");
+    assert_eq!(entry["UNDERSCORE_FOO"], "foo");
+    assert_eq!(entry["SPAM_WITH_EGGS"], "false");
 }